+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the "bpmn-sketch-miner" extension will be documented in this file.
 
+## [1.1.0]
+
+- Refresh the web view on save
+- Adds AutoRefresh option 
+    - Adds a `AutoRefresh` configuration option to automatically refresh the web view after changes. 
+    - Adds a `AutoRefreshTimeout` configuration option to adjust the timeout before AutoRefresh is triggered
+- Update the used packages
+
 ## [1.0.5]
 
 - Fix #3: "Command not found"